Added support for units, multipliers and instances
This is a huge overhaul of the plotjuggler-apbin-plugin with
many code changes!

plotjuggler-apbin-plugin now reads and uses
- units (FMTU and UNIT messages)
- multipliers (FMTU and MULT messages)
and applies them correctly to the message data.

The plugin handles instances defined by the unit
character '#'. Instances are now displayed correctly in PlotJuggler.

The plugin tries to convert the time base to unix time
(based on the logged data from GNSS). This is useful if you want
to analyze rosbags and ArduPilot logfiles in parallel.